Refactored to pull json_tranform into its own module
