+++ libbeat/common/jsontransform/transform.go
@@ -1,26 +1,28 @@
-package common
+package jsontransform
 
 import (
 	"encoding/json"
+
+	"github.com/elastic/beats/libbeat/common"
 )
 
-// transformNumbersDict walks a json decoded tree an replaces json.Number
+// TransformNumbers walks a json decoded tree an replaces json.Number
 // with int64, float64, or string, in this order of preference (i.e. if it
 // parses as an int, use int. if it parses as a float, use float. etc).
-func TransformNumbersDict(dict MapStr) {
+func TransformNumbers(dict common.MapStr) {
 	for k, v := range dict {
 		switch vv := v.(type) {
 		case json.Number:
-			dict[k] = TransformNumber(vv)
+			dict[k] = transformNumber(vv)
 		case map[string]interface{}:
-			TransformNumbersDict(vv)
+			TransformNumbers(vv)
 		case []interface{}:
-			TransformNumbersArray(vv)
+			transformNumbersArray(vv)
 		}
 	}
 }
 
-func TransformNumber(value json.Number) interface{} {
+func transformNumber(value json.Number) interface{} {
 	i64, err := value.Int64()
 	if err == nil {
 		return i64
@@ -32,15 +34,15 @@ func TransformNumber(value json.Number) interface{} {
 	return value.String()
 }
 
-func TransformNumbersArray(arr []interface{}) {
+func transformNumbersArray(arr []interface{}) {
 	for i, v := range arr {
 		switch vv := v.(type) {
 		case json.Number:
-			arr[i] = TransformNumber(vv)
+			arr[i] = transformNumber(vv)
 		case map[string]interface{}:
-			TransformNumbersDict(vv)
+			TransformNumbers(vv)
 		case []interface{}:
-			TransformNumbersArray(vv)
+			transformNumbersArray(vv)
 		}
 	}
 }
